fix pedantic clippy warnings
since the project has moved to using a workspace configuration
in cargo, the strict and pedantic linter settings where
not applied to all crates anymore.
To enable linting in the workspace a

```
[lints]
workspace = true
```

block is needed in the Cargo.toml.
This change introduces the necessary configuration and
fixes the resulting linter warnings.

Signed-off-by: Alexander Mohr <alexander.m.mohr@mercedes-benz.com>