allow extending locks by second put call
to simplify locking, it is important that a second call to
put is extending the lock
This behaviour was broken because the UDS ECU tried to
spawn a new tester present task, when one was already running.
This change adds a check to make sure that tester present is only
started when neccessary.

Signed-off-by: Alexander Mohr <alexander.m.mohr@mercedes-benz.com>