chore: remove hashbrown
using std::collections::HashMap<K, V, foldhash::fast::RandomState>
instead, which is virtually the same as using hashbrown.

The advantage is that we do not have to patch schemars and aide
anymore to support hashbrown.
As the only difference between std::collection and hashbrown
is the hasher and we're using the same hasher globally there is no
performance difference.
Even using the default std::hash::RandomState hasher the performance
difference was very minor.

HashSet and HashMap are implemented as type alias in cda-interfaces
which should be used everywhere moving forward.
Should the need arise to change the implementation again to hashbrown
or something else, it can be done by changing 4 lines of code.

Signed-off-by: Alexander Mohr <alexander.m.mohr@mercedes-benz.com>